+++ judge-impl/src/main/java/com/dwarfeng/judge/impl/bean/entity/HibernateSinkerVariable.java
@@ -5,7 +5,6 @@ import com.dwarfeng.judge.sdk.util.Constraints;
 import com.dwarfeng.subgrade.stack.bean.Bean;
 
 import javax.persistence.*;
-import java.util.Date;
 import java.util.Objects;
 
 @Entity
@@ -13,7 +12,7 @@ import java.util.Objects;
 @Table(name = "tbl_sinker_variable")
 public class HibernateSinkerVariable implements Bean {
 
-    private static final long serialVersionUID = -9063036159511983050L;
+    private static final long serialVersionUID = -1651382760273589063L;
 
     // -----------------------------------------------------------主键-----------------------------------------------------------
     @Id
@@ -25,24 +24,8 @@ public class HibernateSinkerVariable implements Bean {
     private String variableStringId;
 
     // -----------------------------------------------------------主属性字段-----------------------------------------------------------
-    @Column(name = "value_type", nullable = false)
-    private int valueType;
-
-    @Column(name = "string_value", columnDefinition = "TEXT")
-    private String stringValue;
-
-    @Column(name = "long_value")
-    private Long longValue;
-
-    @Column(name = "double_value")
-    private Double doubleValue;
-
-    @Column(name = "boolean_value")
-    private Boolean booleanValue;
-
-    @Column(name = "date_value")
-    @Temporal(TemporalType.TIMESTAMP)
-    private Date dateValue;
+    @Column(name = "value", columnDefinition = "TEXT")
+    private String value;
 
     // -----------------------------------------------------------多对一-----------------------------------------------------------
     @ManyToOne(targetEntity = HibernateSinkerInfo.class)
@@ -89,52 +72,12 @@ public class HibernateSinkerVariable implements Bean {
         this.variableStringId = variableStringId;
     }
 
-    public int getValueType() {
-        return valueType;
+    public String getValue() {
+        return value;
     }
 
-    public void setValueType(int valueType) {
-        this.valueType = valueType;
-    }
-
-    public String getStringValue() {
-        return stringValue;
-    }
-
-    public void setStringValue(String stringValue) {
-        this.stringValue = stringValue;
-    }
-
-    public Long getLongValue() {
-        return longValue;
-    }
-
-    public void setLongValue(Long longValue) {
-        this.longValue = longValue;
-    }
-
-    public Double getDoubleValue() {
-        return doubleValue;
-    }
-
-    public void setDoubleValue(Double doubleValue) {
-        this.doubleValue = doubleValue;
-    }
-
-    public Boolean getBooleanValue() {
-        return booleanValue;
-    }
-
-    public void setBooleanValue(Boolean booleanValue) {
-        this.booleanValue = booleanValue;
-    }
-
-    public Date getDateValue() {
-        return dateValue;
-    }
-
-    public void setDateValue(Date dateValue) {
-        this.dateValue = dateValue;
+    public void setValue(String value) {
+        this.value = value;
     }
 
     public HibernateSinkerInfo getSinkerInfo() {
@@ -150,12 +93,7 @@ public class HibernateSinkerVariable implements Bean {
         return getClass().getSimpleName() + "(" +
                 "sinkerLongId = " + sinkerLongId + ", " +
                 "variableStringId = " + variableStringId + ", " +
-                "valueType = " + valueType + ", " +
-                "stringValue = " + stringValue + ", " +
-                "longValue = " + longValue + ", " +
-                "doubleValue = " + doubleValue + ", " +
-                "booleanValue = " + booleanValue + ", " +
-                "dateValue = " + dateValue + ", " +
+                "value = " + value + ", " +
                 "sinkerInfo = " + sinkerInfo + ")";
     }
 }